requirements.txt
click==8.1.3
GitPython==3.1.11
graphviz==0.20
icecream==2.1.2
jsonschema==4.6.1
networkx==2.8
nltk==3.7
numpy==1.19.5
pandarallel==1.6.1
pandas==1.4.2
PyDriller==2.1
rich==12.3.0
spacy==3.3.0
stanfordcorenlp==3.9.1.1
textacy==0.12.0
timeout_decorator==0.5.0
tree_sitter==0.20.0
typing_extensions==4.4.0
PyYAML==5.4.1
pygraphviz==1.9